requirements.txt
Package          Version            
---------------- -------------------
absl-py          0.9.0              
certifi          2019.11.28         
chardet          3.0.4              
colorama         0.4.3              
cycler           0.10.0             
deepdiff         4.0.9              
docopt           0.6.2              
enum34           1.1.6              
future           0.18.2             
gitdb2           2.0.6              
GitPython        3.0.5              
idna             2.8                
jsonpickle       1.2                
kiwisolver       1.1.0              
matplotlib       3.1.2              
mkl-fft          1.0.15             
mkl-random       1.1.0              
mkl-service      2.3.0              
mock             3.0.5              
mpyq             0.2.5              
munch            2.5.0              
numpy            1.18.1             
olefile          0.46               
ordered-set      3.1.1              
packaging        20.0               
Pillow           7.0.0              
pip              19.3.1             
portpicker       1.3.1              
protobuf         3.11.2             
py-cpuinfo       5.0.0              
pygame           1.9.6              
pyparsing        2.4.6              
PySC2            3.0.0              
python-dateutil  2.8.1              
PyYAML           5.3                
requests         2.22.0             
s2clientprotocol 4.11.3.77661.0     
s2protocol       4.11.3.77661.0     
sacred           0.8.1              
scipy            1.4.1              
setuptools       44.0.0.post20200106
six              1.13.0             
sk-video         1.1.10             
SMAC             0.1.0b1            
smmap2           2.0.5              
torch            1.4.0              
torchvision      0.5.0              
urllib3          1.25.7             
websocket-client 0.57.0             
wheel            0.33.6             
whichcraft       0.6.1              
wrapt            1.11.2             
